change gmaps at chart to component
