+++ resources/js/Pages/Chart/ChartVIF.tsx
@@ -20,6 +20,7 @@ import AggregationRST from "@/Components/Chart/AggregationRST";
 import AggregationSingle from "@/Components/Chart/AggregationSingle";
 import AggregationRSTIN from "@/Components/Chart/AggregationRSTIN";
 import calculateMetrics from "@/helpers/analysis/calculate-metric";
+import GoogleMap from "@/Components/Map/GoogleMap";
 
 export default function ChartVIF({
                                      trafo,
@@ -70,25 +71,6 @@ export default function ChartVIF({
 
     const frequencyAggregation = calculateMetrics(frequencies.map(frequency => frequency.frequency_r));
     const { frequency_r = 0 } = frequencies[frequencies.length - 1] || {};
-
-    const renderMarker = (map: any, maps: any) => {
-        return new maps.Marker({
-            position: {
-                lat: Number(trafo.latitude),
-                lng: Number(trafo.longitude),
-            },
-            map,
-            title: 'test marker'
-        });
-    }
-
-    const defaultProps = {
-        center: {
-            lat: Number(trafo.latitude),
-            lng: Number(trafo.longitude),
-        },
-        zoom: 15,
-    }
 
     return (
         <>
@@ -186,15 +168,12 @@ export default function ChartVIF({
                         </Box>
                     </Grid>
                     <Grid item xs={12} md={4}>
-                        <Box sx={{ height: '75vh', width: '100%' }}>
-                            <GoogleMapReact
-                                bootstrapURLKeys={{ key: mapApiKey }}
-                                defaultCenter={defaultProps.center}
-                                defaultZoom={defaultProps.zoom}
-                                yesIWantToUseGoogleMapApiInternals
-                                onGoogleApiLoaded={({ map, maps }) => renderMarker(map, maps)}
-                            />
-                        </Box>
+                        <GoogleMap
+                            lat={Number(trafo.latitude)}
+                            lng={Number(trafo.longitude)}
+                            title={trafo.name}
+                            height={'700px'}
+                        />
                     </Grid>
                 </Grid>
             </Container>
